+++ src/main/java/dyvil/lang/Int.java
@@ -30,25 +30,25 @@ public abstract class Int implements Number
 	@Override
 	public int intValue()
 	{
-		return (int) this.value;
+		return this.value;
 	}
 	
 	@Override
 	public long longValue()
 	{
-		return (long) this.value;
+		return this.value;
 	}
 	
 	@Override
 	public float floatValue()
 	{
-		return (float) this.value;
+		return this.value;
 	}
 	
 	@Override
 	public double doubleValue()
 	{
-		return (double) this.value;
+		return this.value;
 	}
 	
 	// Unary operators
